requirements.txt
numpy
pandas
scikit-learn
tensorflow
pyyaml
joblib
streamlit
